Chapters/Chapter 13/mlflow/data_quality_agent/pyproject.toml
[project]
name = "data_quality_agent"
version = "0.0.1"
authors = [{ name = "alan.dennis@covasant.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "mlflow",
    "PyYAML",
    "pandas"
]

[dependency-groups]
dev = [
    "pytest",
    "databricks-dlt",
    "databricks-connect>=15.6,<15.7",
]

[project.scripts]
main = "data_quality_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 125
